% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_std_plot_plot}
\alias{add_std_plot_plot}
\title{Adds standard color scheme and geoms to a ggplot object}
\usage{
add_std_plot_plot(..., max_n = 16000, step_x_break = 1000)
}
\arguments{
\item{...}{Incoming plot}

\item{max_n}{Last break on x axis}

\item{step_x_break}{Steps between x axis ticks}
}
\description{
Adds standard color scheme and geoms to a ggplot object
}
