% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_energy.R
\name{clean_energy_measurement}
\alias{clean_energy_measurement}
\title{Transform raw energy measurements into a tidy tibble for further analysis}
\usage{
clean_energy_measurement(df, df_time)
}
\arguments{
\item{df}{Data frame containing raw energy data from measurement script}

\item{df_time}{Data frame with processed time data for energy calculation}
}
\description{
Transform raw energy measurements into a tidy tibble for further analysis
}
