% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_energy.R
\name{clean_energy_mm_old}
\alias{clean_energy_mm_old}
\title{Transform raw energy measurements into a tidy tibble for further analysis}
\usage{
clean_energy_mm_old(df, time_df, idle_draw)
}
\arguments{
\item{df}{Data frame containing raw energy data from measurement script}

\item{idle_draw}{Idle power draw of the gpu}

\item{df_time}{Data frame with processed time data for energy calculation}
}
\description{
Depreciated function that is only used to transform older raw data. The
measurement script was changed relatively early. However, some measurements
may still depend on this function.
}
