% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_measurement}
\alias{clean_measurement}
\title{Goes through the whole cleaning process for time as well as energy}
\usage{
clean_measurement(time_path, energy_path)
}
\arguments{
\item{time_path}{Path of raw time measurements}

\item{energy_path}{Path of raw energy measurements}
}
\description{
Goes through the whole cleaning process for time as well as energy
}
