% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_time.R
\name{clean_time_m_old}
\alias{clean_time_m_old}
\title{Transform raw time measurements into a tidy tibble for further analysis}
\usage{
clean_time_m_old(df)
}
\arguments{
\item{df}{Data frame containing raw time data from measurement script}
}
\description{
Depreciated function that is only used to transform older raw data. The
measurement script was changed relatively early. However, some measurements
may still depend on this function.
}
