% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_time.R
\name{clean_time_measurement}
\alias{clean_time_measurement}
\title{Transform raw time measurements into a tidy tibble for further analysis}
\usage{
clean_time_measurement(df)
}
\arguments{
\item{df}{Data frame containing raw time data from measurement script}
}
\description{
Transform raw time measurements into a tidy tibble for further analysis
}
