% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_std_measurement}
\alias{plot_std_measurement}
\title{Plots a standard measurement using ggplot2's ggplot}
\usage{
plot_std_measurement(data, metric, labels = scales::comma, ...)
}
\arguments{
\item{data}{Data that is vizualized}

\item{metric}{Metric displayed on y axis}

\item{labels}{Labels of y axis}

\item{...}{Passes additional arguments through ggplot}
}
\description{
Plots a standard measurement using ggplot2's ggplot
}
